adds a basic .tmux.conf
These are the basic settings I needed to get tmux working the way I
wanted.  My original goal was to be able to use VI commands on the
terminal, and tmux seems like the best way to do it, and it comes with
the benifit of using multiplexed screens, like I should be in the habit
of.